TST: Move misc old tests in test_examples into newer test modules.
